change variables type

+++ src/DO.h
@@ -4,7 +4,7 @@
 
 class DO{
     public:
-        DO(uint8_t pin, uint32_t vref, uint32_t aref, int cal_mode);
+        DO(uint8_t pin, double vref, double aref, int cal_mode);
         ~DO();
         void setTemperature(float temp);
         float getTemperature();
@@ -13,7 +13,7 @@ class DO{
         float getDOValue();
         void outputDOData(int delay_time);
     private:
-        uint32_t _vref, _aref;
+        double _vref, _aref;
         float _temp;
         uint8_t _pin;
         int _cal_mode, _delay_time;
